+++ piwigo/Album/Discover/SearchImagesViewController.m
@@ -459,7 +459,7 @@
                 [self.imagesCollection reloadItemsAtIndexPaths:indexPaths];
             });
             // Should we continue loading images?
-            NSLog(@"==> Should we continue loading images? (scrolled to %ld)", self.didScrollToImageIndex);
+            NSLog(@"==> Should we continue loading images? (scrolled to %ld)", (long)self.didScrollToImageIndex);
             if (self.didScrollToImageIndex >= downloadedImageCount+imagesPerPage) {
                 [self needToLoadMoreImages];
             }
